Make config optional. Move all defaults to 1 location
